% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandidateHyperparamatersRF.R
\name{CandidateHyperparamatersRF}
\alias{CandidateHyperparamatersRF}
\title{Candidate Hyperparameters - RF}
\usage{
CandidateHyperparamatersRF(numr,mtry_min,mtry_max,numtrees_min,numtrees_max,samplefraction_min,samplefraction_max,minnodesize_min,minnodesize_max)
}
\arguments{
\item{numr}{The number of hyperparamter values to generate}

\item{mtry_min}{min mtry hyperparameter value}

\item{mtry_max}{max mtry hyperparameter value}

\item{numtrees_min}{min numtrees hyperparameter value}

\item{numtrees_max}{max numtrees hyperparameter value}

\item{samplefraction_min}{min samplefraction hyperparameter value}

\item{samplefraction_max}{max samplefraction hyperparameter value}

\item{minnodesize_min}{min minnodesize hyperparameter value}

\item{minnodesize_max}{max minnodesize hyperparameter value}
}
\value{
Dataframe of \code{numr} entries for candidate hyperparameters to use with random forest
}
\description{
Generate candidate hyperparameter set for Random Forest to use with new dataset
}
\examples{
CandidateHyperparamatersRF(10, 0.3, 0.8, 50, 100, 0.4, 1, 0, 1)
CandidateHyperparamatersRF(20, mtry_min=0.6, mtry_max=0.9, samplefraction_min=0.7, minnodesize_min=0, minnodesize_max=1)
}
