% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandidateHyperparametersXGB.R
\name{CandidateHyperparamatersXGB}
\alias{CandidateHyperparamatersXGB}
\title{Candidate Hyperparameters - XGB}
\usage{
CandidateHyperparamatersXGB(numr,nrounds_min,nrounds_max,eta_min,eta_max,subsample_min,subsample_max,max_depth_min,max_depth_max,min_child_weight_min,min_child_weight_max,colsample_bytree_min,colsample_bytree_max,lambda_min,lambda_max,alpha_min,alpha_max)
}
\arguments{
\item{numr}{The number of hyperparamter values to generate}

\item{nrounds_min}{min mtry hyperparameter value}

\item{nrounds_max}{max mtry hyperparameter value}

\item{eta_min}{min eta hyperparameter value}

\item{eta_max}{max eta hyperparameter value}

\item{subsample_min}{min subsample hyperparameter value}

\item{subsample_max}{max subsample hyperparameter value}

\item{max_depth_min}{min max_depth hyperparameter value}

\item{max_depth_max}{max max_depth hyperparameter value}

\item{min_child_weight_min}{min min_child_weight hyperparameter value}

\item{min_child_weight_max}{max min_child_weight hyperparameter value}

\item{colsample_bytree_min}{min colsample_bytree hyperparameter value}

\item{colsample_bytree_max}{max colsample_bytree hyperparameter value}

\item{lambda_min}{min lambda hyperparameter value}

\item{lambda_max}{max lambda hyperparameter value}

\item{alpha_min}{min alpha hyperparameter value}

\item{alpha_max}{max alpha hyperparameter value}
}
\value{
Dataframe of \code{numr} entries for candidate hyperparameters to use with xgboost
}
\description{
Generate candidate hyperparameter set for XGBoost to use with new dataset
}
\examples{
CandidateHyperparamatersXGB(100, 50, 200, 0.05, 0.2, 0.6, 0.7, 6, 10, 0, 1, 0.75, 0.8, 0, 0.5, 0.2, 1)
}
