\name{ExtractMetafeatures}
\alias{ExtractMetafeatures}
\title{Extract Metafeatures}
\usage{
ExtractMetafeatures(X_train,y_train)

}
\description{
Extracts the relevant metafeatures from the user's data to match the metadata used with the package's metalearners.
}
\examples{
ExtractMetafeatures(X_train,y_train)
}
