% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMetalearners.R
\name{GenerateMetalearners}
\alias{GenerateMetalearners}
\title{Generate Metalearners}
\usage{
GenerateMetalearners(obj_fn,ml_algo)
}
\arguments{
\item{obj_fn}{The objective function used to evaluate models (KGE/NSE)}

\item{ml_algo}{The type of machine learning algorithm to train (RF/XGB)}
}
\value{
will save the specified model as an .rds file for future use/reference
}
\description{
Train metalearners for specific objective function and ML algorithm
}
\examples{
GenerateMetalearners("KGE","RF")
GenerateMetalearners("NSE","XGB")
}
