% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadMetalearners.R
\name{LoadMetalearner}
\alias{LoadMetalearner}
\title{Load Metalearners}
\usage{
LoadMetalearner(obj_fn,ml_algo)
}
\arguments{
\item{obj_fn}{The objective function used to evaluate models (KGE/NSE)}

\item{ml_algo}{The type of machine learning algorithm to train (RF/XGB)}
}
\value{
will save the specified model as an .rds file for future use/reference
}
\description{
Load existing metalearners based on ML algorithm and objective function used to evaluate it. This metalearner can be used with unseen data.
}
\examples{
LoadMetalearner("KGE","RF")
LoadMetalearner("NSE","XGB")
}
